requirements.txt
# Core dependencies
web3==6.15.1
eth-account==0.11.0
cryptography==42.0.2
python-dotenv==1.0.1

# Notion API
notion-client==2.2.1

# HTTP requests
requests==2.31.0
aiohttp==3.9.3

# Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Database
sqlalchemy==2.0.25

# Utilities
pydantic==2.6.1
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
